Backport merge scope fix from Rails 4.1
The commit that fixed the issue on Rails 4.0 was:
https://github.com/rails/rails/commit/c8d889905dab071f1d8a166b25fa69cdd31dc176

However this backport the most updated version from latest 4-1-stable.

The idea behind this patch is:
Before if we had a codition like this:
Left side:
where(id: 1).where(id: 2)

Right side:
where(foo: 'bar')

A scope.merge of both sides would colapse the two ids on the left.
This is incorrect, as a merge should only merge right into left.

As stated this is the current behaviour on modern Rails versions.